+++ nafas/params.py
@@ -101,7 +101,7 @@ STANDARD_MENU = {
         17: "Coherent",
         18: "Fire",
         19: "Retention",
-        20: "Swooning Breath"},
+        20: "Swooning"},
     "level": {
         1: "Beginner",
         2: "Medium",
@@ -165,11 +165,11 @@ RETENTION_MEDIUM = {"ratio": [2, 0, 2, 12], "unit": 1, "pre": 3, "cycle": 10}
 
 RETENTION_ADVANCED = {"ratio": [2, 0, 2, 12], "unit": 1, "pre": 3, "cycle": 15}
 
-SWOONING_BREATH_BEGINNER = {"ratio": [5, 10, 7, 0], "unit": 1, "pre": 3, "cycle": 7}
+SWOONING_BEGINNER = {"ratio": [5, 10, 7, 0], "unit": 1, "pre": 3, "cycle": 7}
 
-SWOONING_BREATH_MEDIUM = {"ratio": [5, 10, 7, 0], "unit": 1, "pre": 3, "cycle": 12}
+SWOONING_MEDIUM = {"ratio": [5, 10, 7, 0], "unit": 1, "pre": 3, "cycle": 12}
 
-SWOONING_BREATH_ADVANCED = {"ratio": [5, 10, 7, 0], "unit": 1, "pre": 3, "cycle": 21}
+SWOONING_ADVANCED = {"ratio": [5, 10, 7, 0], "unit": 1, "pre": 3, "cycle": 21}
 
 
 RELAX2_BEGINNER = {
@@ -537,9 +537,9 @@ PROGRAMS = {"Clear Mind": {"Beginner": CLEAR_MIND_BEGINNER,
             "Retention": {"Beginner": RETENTION_BEGINNER,
                                  "Medium": RETENTION_MEDIUM,
                                  "Advanced": RETENTION_ADVANCED},
-            "Swooning Breath": {"Beginner": SWOONING_BREATH_BEGINNER,
-                                "Medium": SWOONING_BREATH_MEDIUM,
-                                "Advanced": SWOONING_BREATH_ADVANCED},
+            "Swooning": {"Beginner": SWOONING_BEGINNER,
+                                "Medium": SWOONING_MEDIUM,
+                                "Advanced": SWOONING_ADVANCED},
             }
 
 SPEAKER_LIST = [